string_generator/string_gen/serializers.py
from rest_framework import serializers
from .models import StringGenerator


class StringGeneratorSerializer(serializers.ModelSerializer):

    class Meta:
        model = StringGenerator
        fields = ('template', 'par')
