requirements.txt
Django==1.11.4
djangorestframework==3.6.4
pkg-resources==0.0.0
psycopg2==2.7.1
